Parameters to configure the test

+++ src/main/java/com/migangqui/example/postgresql/bootstrap/InitComponent.java
@@ -2,7 +2,10 @@ package com.migangqui.example.postgresql.bootstrap;
 
 import java.util.List;
 
+import javax.annotation.PostConstruct;
+
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.core.env.Environment;
 import org.springframework.stereotype.Component;
 
 import com.migangqui.example.postgresql.entity.Event;
@@ -19,61 +22,79 @@ public class InitComponent {
 
 	@Autowired
 	private EventRepository eventRepository;
-	
+
+	@Autowired
+	private Environment env;
+
 	private GeometryFactory geometryFactory = new GeometryFactory();
-	
-	
 
-//	@PostConstruct
+	@PostConstruct
 	public void init() {
-		eventRepository.deleteAll();
+		boolean insertData = env.getProperty("data.insert", boolean.class);
 
-		log.info("Total object in db {}", eventRepository.count());
-		Event event;
 		long startTime = System.currentTimeMillis();
-		for (int i = 0; i < 250_000; i++) {
 
-			log.info("Registering event A and B {}", i+1);
+		if (insertData) {
 
-			event = new Event();
-			event.setName("eventA"+i);
-			event.setLocation(geometryFactory.createPoint(new Coordinate(-5.994724, 37.397159)));
+			log.info("Inserting data");
+			
+			if (env.getProperty("data.delete", boolean.class)) {
+				eventRepository.deleteAll();
+			}
 
-			eventRepository.save(event);
+			log.info("Total object in db {}", eventRepository.count());
+			Event event;
+			
+			int quantityTotal = env.getProperty("data.quantity.total", int.class);
+			int quantityFind = env.getProperty("data.quantity.find", int.class);
+			
+			for (int i = 0; i < quantityTotal; i++) {
 
-			event = new Event();
-			event.setName("eventB"+i);
-			event.setLocation(geometryFactory.createPoint(new Coordinate(-4.773935, 37.882503)));
+				log.info("Registering event A and B {}", i + 1);
 
-			eventRepository.save(event);
+				event = new Event();
+				event.setName("eventA" + i);
+				event.setLocation(geometryFactory.createPoint(new Coordinate(-5.994724, 37.397159)));
+
+				eventRepository.save(event);
+
+				event = new Event();
+				event.setName("eventB" + i);
+				event.setLocation(geometryFactory.createPoint(new Coordinate(-4.773935, 37.882503)));
+
+				eventRepository.save(event);
+
+			}
+
+			for (int i = 0; i < quantityFind; i++) {
+
+				log.info("Registering event C {}", i + 1);
+
+				event = new Event();
+				event.setName("eventC" + i);
+				event.setLocation(geometryFactory.createPoint(new Coordinate(2.325399, 48.823206)));
+
+				eventRepository.save(event);
+
+			}
+
+			log.info("Time in insert data in DB -> {}", System.currentTimeMillis() - startTime);
 
 		}
-
-		for (int i = 0; i < 123; i++) {
-
-			log.info("Registering event C {}", i+1);
-
-			event = new Event();
-			event.setName("eventC"+i);
-			event.setLocation(geometryFactory.createPoint(new Coordinate(2.325399, 48.823206)));
-
-			eventRepository.save(event);
-
-		}
-
-		log.info("Time in insert data in DB -> {}", System.currentTimeMillis() - startTime);
 
 		GeometricShapeFactory shapeFactory = new GeometricShapeFactory();
-	    shapeFactory.setNumPoints(32);
-	    shapeFactory.setCentre(new Coordinate(2.320237, 48.821198));
-	    shapeFactory.setSize(1);
-	    shapeFactory.createCircle();
+		shapeFactory.setNumPoints(32);
+		shapeFactory.setCentre(new Coordinate(2.320237, 48.821198));
+		shapeFactory.setSize(1);
+		shapeFactory.createCircle();
 
-	    startTime = System.currentTimeMillis();
-	    List<Event> events = eventRepository.findByLocationWithIn(shapeFactory.createCircle());
-		log.info("Time in find {} object betweewn {} -> {}", events.size(), eventRepository.count(), System.currentTimeMillis() - startTime);
+		startTime = System.currentTimeMillis();
+		List<Event> events = eventRepository.findByLocationWithIn(shapeFactory.createCircle());
+		log.info("Time in find {} object betweewn {} -> {} ms", events.size(), eventRepository.count(),
+				System.currentTimeMillis() - startTime);
 
 		System.exit(0);
+
 	}
-	
+
 }
